changeinput

+++ src/@StochasticSystem/monte_carlo_average.m
@@ -1,23 +1,32 @@
 function [w,PSD_a]=monte_carlo_average(obj,method,PSDpair,nRealization,clusterRun)
-assert(nRealization > 1, ' Please use sde.solver for one simulation')
 
-if clusterRun
-    euler = parcluster('local');
-    pool = parpool(euler,24);
+if nRealization>1
+
+    if clusterRun
+        euler = parcluster('local');
+        pool = parpool(euler,24);
+    else
+        pool = parpool('local',2);
+    end
+
+    [w,X] = obj.sde_solver(method,PSDpair);
+
+    parfor i=1:nRealization-1
+        [~,outputPSD] = obj.sde_solver(method,PSDpair);
+        X = X+outputPSD;
+    end
+    PSD_a = X/nRealization;
+
+
+    pool.delete()
 else
-    pool = parpool('local',2);
+    
+    [w,PSD_a]=obj.sde_solver(method,PSDpair);
+    
+    
 end
 
-[w,X] = obj.sde_solver(method,PSDpair);
 
-parfor i=1:nRealization-1
-    [~,outputPSD] = obj.sde_solver(method,PSDpair);
-    X = X+outputPSD;
-end
-PSD_a = X/nRealization;
-
-
-pool.delete()
 
 end
